third_party/WebKit/LayoutTests/bluetooth/disconnect/disconnect-twice-in-a-row.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => {
      gattServer.disconnect();
      assert_false(gattServer.connected);
      gattServer.disconnect();
      assert_false(gattServer.connected);
    });
}, 'Calling disconnect twice in a row still results in \'connected\' ' +
   'being false.');
</script>
